Safari-609.1.11.1/LayoutTests/platform/wpe/crypto/subtle/ecdh-import-spki-key-ecdh-identifier-expected.txt

Test importing P-256, P-384 and P-521 SPKI ECDH keys that are using the ecDH algorithm identifier

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Importing a P-256 key that uses the ecDH algorithm identifier ...
PASS publicKey.toString() is '[object CryptoKey]'
PASS publicKey.type is 'public'
PASS publicKey.extractable is true
PASS publicKey.algorithm.name is 'ECDH'
PASS publicKey.algorithm.namedCurve is 'P-256'
PASS publicKey.usages is [ ]
Importing a P-384 key that uses the ecDH algorithm identifier ...
PASS publicKey.toString() is '[object CryptoKey]'
PASS publicKey.type is 'public'
PASS publicKey.extractable is true
PASS publicKey.algorithm.name is 'ECDH'
PASS publicKey.algorithm.namedCurve is 'P-384'
PASS publicKey.usages is [ ]
Importing a P-521 key that uses the ecDH algorithm identifier ...
PASS publicKey.toString() is '[object CryptoKey]'
PASS publicKey.type is 'public'
PASS publicKey.extractable is true
PASS publicKey.algorithm.name is 'ECDH'
PASS publicKey.algorithm.namedCurve is 'P-521'
PASS publicKey.usages is [ ]
PASS successfullyParsed is true

TEST COMPLETE
